PDF-Review/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pdfplumber==0.10.3
openai==1.12.0
redis==5.0.1
python-dotenv==1.0.0
Werkzeug==3.0.1
